// frontend/src/api/schema.d.ts
export interface components {
  schemas: {
// ...
    CatalogNews: {
      /**
       * At
       * Format: date-time
       */
      at: string;
      /**
       * Kind
       * @description new новое решение или выгрузка, spec характеристика, price цена, task решение пошло в подбор
       * @enum {string}
       */
      kind: "new" | "spec" | "price" | "task";
      /** Title */
      title: string;
      /** Text */
      text: string;
      /** Solution Id */
      solution_id?: string | null;
      /** Photo Url */
      photo_url?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}